+++ src/components/chat/chatScreen/imageChatLogic.js
