--- src/components/ui/NewChatIcon.js
@@ -1,16 +1,0 @@
-import React, {Component} from 'react'
-import {View, Text} from 'react-native'
-
-const NewChatIcon = () => {
-
-  return (
-    <View>
-      <Text>
-        Not Seen
-      </Text>
-    </View>
-  )
-
-}
-
-export default NewChatIcon
